% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch_opt.R
\name{alpha2N}
\alias{alpha2N}
\title{Internal function. convert mixture proportions to mixture sample size with a fixed total sample size}
\usage{
alpha2N(n, alpha)
}
\arguments{
\item{n}{total number of samples}

\item{alpha}{the vector of mixture proportions}
}
\value{
the vector of sample sizes for each mixture component
}
\description{
Internal function. convert mixture proportions to mixture sample size with a fixed total sample size
}

