% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch_opt.R
\name{do.mixture.sample}
\alias{do.mixture.sample}
\title{Internal function. sample from the mixture distribution \eqn{q_{\alpha}}}
\usage{
do.mixture.sample(seed, b, n, J, mixture.param, alpha, fname, rpname, rqname,
  dpname, dqname)
}
\arguments{
\item{seed}{seed for sampling}

\item{b}{batch index for the samples}

\item{n}{total number of samples}

\item{J}{number of mixture components, including the defensive one}

\item{mixture.param}{mixture.param = list(p, J, ...), where \eqn{p} is the dimension of the sample, and \eqn{J} is the number of mixture components, including the defensive one. mixture.param should be compatible with user defined functions \emph{f(n, j, mixture.param), rp(n, mixture.param), rq(n, j, mixture.param), dp(xmat, mixture.param), dq(xmat, j, mixture.param)}}

\item{alpha}{vector of mixture proportions}

\item{fname}{name of user defined function \emph{fname(xmat, j, mixture.param)}. \emph{xmat} is an \eqn{n \times p} matrix of \eqn{n} samples with \eqn{p} dimensions. \emph{fname} returns a vector of function values for each row in \emph{xmat}. \emph{fname} is defined for \eqn{j = 1, \cdots, J}. \eqn{j = 1, \cdots, J - 1} corresponds to different proposal mixture components, and \eqn{j = J} corresponds to the defensive mixture component.}

\item{rpname}{name of user definded function \emph{rpname(n, mixture.param)}. It generates \eqn{n} random samples from target distribution \emph{pname}. Parameters can be specified in \emph{mixture.param}. \emph{rpname} returns an \eqn{n \times p} matrix.}

\item{rqname}{name of user defined function \emph{rqname(n, j, mixture.param)}. It generate \eqn{n} random samples from the \eqn{j^{th}} mixture component of proposal mixture distribution. \emph{rqname} returns an \eqn{n \times p} matrix. \emph{rqname} is defined for \eqn{j = 1, \cdots, J - 1}.}

\item{dpname}{name of user defined function \emph{dpname(xmat, mixture.param)}. It returns the density of \emph{xmat} from the target distribution \eqn{pname} as a vector of length \emph{nrow(xmat)}. Note that only the ratio between \emph{dpname} and \emph{dqname} matters. So \emph{dpname} and \emph{dqname} can return values of \eqn{C \times}\emph{dpname} and \eqn{C \times}\emph{dqname} respectively.}

\item{dqname}{name of user defined function \emph{dqname(xmat, j, mixture.param)}. It returns the density of \emph{xmat} from the proposal distribution \eqn{q_j} as a vector of length \emph{nrow(xmat)}. \emph{dqname} is defined for \eqn{j = 1, \cdots, J - 1}. Note that only the ratio between \emph{dpname} and \emph{dqname} matters. So \emph{dpname} and \emph{dqname} can  return values of \eqn{C \times}\emph{dpname} and \eqn{C \times}\emph{dqname} respectively.}
}
\value{
a list of \describe{
  \item{x}{the matrix(size \eqn{n \times p}) of samples from the mixture distribution \eqn{q_{\alpha}}}
  \item{fx}{the vector of \emph{fname} evaluated at sample matrix \emph{x}}
  \item{qx}{the matrix of densities under each mixture component, i.e. \emph{qx[,j]} is the density under \eqn{q_j}. The defensive component is  the \eqn{J^{th}} column}
  \item{px}{the vector of densities under proposal distribution \emph{pname}}
  \item{alpha}{the rounded \emph{alpha} used for sampling, i.e. \eqn{alpha2N(n, alpha)/n}}
}
}
\description{
Internal function. sample from the mixture distribution \eqn{q_{\alpha}}
}

