% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch_opt.R
\name{get.index.b}
\alias{get.index.b}
\title{Internal function. Get the row index in the stacked sample matrices for the \eqn{b^{th}} batch}
\usage{
get.index.b(batch.size, b)
}
\arguments{
\item{batch.size}{vector of batch sizes}

\item{b}{the index of the batch of interest}
}
\value{
the row index or sample matrices of batch b for all b's in b.vec
}
\description{
Internal function. Get the row index in the stacked sample matrices for the \eqn{b^{th}} batch
}
\details{
In the computation, samples from each batches are stacked be rows to form x.mat,
as well as fx.mat, px.mat, qx.mat. To extract the samples from batch b and their corresponding
fx, px, qx values, the indices for batch b will be needed. They can be obtained by calling
get.index.b(batch.size, b)
}

