% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch_opt.R
\name{get.initial.alpha}
\alias{get.initial.alpha}
\title{Internal function. Calculate the initial alpha vector for the optimization of \emph{alpha} with a lower bound constraint}
\usage{
get.initial.alpha(eps, J)
}
\arguments{
\item{eps}{a vector of lower bound values for vector \emph{alpha}}

\item{J}{number of mixture components}
}
\value{
The initial alpha vector for optimization
}
\description{
Internal function. Calculate the initial alpha vector for the optimization of \emph{alpha} with a lower bound constraint
}

